refactored schema, resolvers, and graphql server
