+++ server/graphql/resolvers/index.js
@@ -1,0 +1,7 @@
+const resolvers = {
+  Query: {
+    info: () => `This is the API for Cody's Untitled Game.`
+  }
+}
+
+module.exports = resolvers;
