+++ server/graphql/server.js
@@ -1,0 +1,9 @@
+const { GraphQLServer, } = require('graphql-yoga');
+const schema = require('./schema');
+
+
+const server = new GraphQLServer({
+  schema
+})
+
+module.exports = server;
