+++ server/tests/server.test.js
@@ -1,6 +1,5 @@
 const { graphql } = require('graphql');
-
-const { schema } = require('../server.js');
+const schema = require('../graphql/schema');
 
 const infoQuery = `
   query{
@@ -9,7 +8,7 @@ const infoQuery = `
 `
 
 describe('GraphQL API', () => {
-  it(`infoQuery returns { "data": { "info": "This is the API for Cody's Untitled Game." } }`, async () =>{
+  it(`should return "This is the API for Cody's Untitled Game." when sent infoQuery`, async () =>{
     const response = await graphql(schema, infoQuery, null, {}, {})
     expect(response.data.info).toBe("This is the API for Cody's Untitled Game.");
   });
